add function to return weight diff; rename formatting.ts -> helpers.ts
